refactor(gate): split index.ts (1206 lines) into handlers/
God file → handlers/{request,review,read,issues,messages}.ts split.
index.ts shrinks to 158 lines and contains only routing + HELP.

- handlers/internal.ts — shared helpers (type C, readStdin,
  loadAllRequestsAsJson, truncateCodePoints, parseOptionalIntOption)
- handlers/request.ts — request lifecycle + list/show + marker helpers
- handlers/review.ts   — gate review
- handlers/read.ts     — voices / tail / whoami / chain
- handlers/issues.ts   — issues subverbs + promote
- handlers/messages.ts — message / broadcast / inbox / mark-read

Behavior unchanged. Existing 127 tests still pass. Manual smoke
runs request → approve → execute → complete → review → show / tail /
chain / pending on a fresh content_root, all verbs green.

formatReviewMarkers and computeReviewMarkerWidth are re-exported from
index.ts for backward-compat with tests/interface/reviewMarkers.test.ts.
