+++ src/interface/gate/handlers/issues.ts
@@ -1,0 +1,123 @@
+import {
+  ParsedArgs,
+  requireOption,
+  optionalOption,
+} from '../../shared/parseArgs.js';
+import { C, truncateCodePoints } from './internal.js';
+
+export async function issuesCmd(c: C, args: ParsedArgs): Promise<number> {
+  const sub = args.positional[0];
+  if (sub === 'promote') {
+    return await issuesPromote(c, args);
+  }
+  if (sub === 'add') {
+    const from = requireOption(args, 'from', '--from required', 'GUILD_ACTOR');
+    const severity = requireOption(args, 'severity', '--severity required');
+    const area = requireOption(args, 'area', '--area required');
+    const text = args.positional.slice(1).join(' ');
+    if (!text) throw new Error('Usage: gate issues add ... <text>');
+    const i = await c.issueUC.add({ from, severity, area, text });
+    process.stdout.write(`✓ issue: ${i.id.value}\n`);
+    return 0;
+  }
+  if (sub === 'list' || sub === undefined) {
+    const state = optionalOption(args, 'state');
+    const items = await c.issueUC.list(state);
+    for (const i of items) {
+      const j = i.toJSON();
+      process.stdout.write(
+        `${j['id']} [${j['severity']}/${j['area']}] ${j['state']} — ${j['text']}\n`,
+      );
+    }
+    return 0;
+  }
+  // State transitions: resolve, defer, start, reopen.
+  const nextState = resolveIssueVerb(sub);
+  if (nextState !== undefined) {
+    const id = args.positional[1];
+    if (!id) throw new Error(`Usage: gate issues ${sub} <id>`);
+    const issue = await c.issueUC.setState(id, nextState);
+    process.stdout.write(`✓ issue ${issue.id.value}: → ${nextState}\n`);
+    return 0;
+  }
+  throw new Error(`unknown issues sub: ${sub}`);
+}
+
+async function issuesPromote(c: C, args: ParsedArgs): Promise<number> {
+  const id = args.positional[1];
+  if (!id) {
+    throw new Error(
+      'Usage: gate issues promote <id> --from <m> [--executor <m>] ' +
+        '[--auto-review <m>] [--action <a>] [--reason <r>]',
+    );
+  }
+  const from = requireOption(args, 'from', '--from required', 'GUILD_ACTOR');
+  const executor = optionalOption(args, 'executor');
+  const autoReview = optionalOption(args, 'auto-review');
+  const actionOverride = optionalOption(args, 'action');
+  const reasonOverride = optionalOption(args, 'reason');
+
+  const issue = await c.issueUC.find(id);
+  if (!issue) {
+    process.stderr.write(`issue not found: ${id}\n`);
+    return 1;
+  }
+  const j = issue.toJSON();
+  if (j['state'] === 'resolved') {
+    throw new Error(
+      `issue ${id} is already resolved; cannot promote a resolved issue`,
+    );
+  }
+
+  const issueText = String(j['text']);
+  const shortText = truncateCodePoints(issueText, 60);
+  const action = actionOverride ?? `Fix issue ${id}: ${shortText}`;
+  const reason =
+    reasonOverride ??
+    `Promoted from ${id} (${j['severity']}/${j['area']}): ${issueText}`;
+
+  const input: Parameters<typeof c.requestUC.create>[0] = {
+    from,
+    action,
+    reason,
+  };
+  if (executor !== undefined) input.executor = executor;
+  if (autoReview !== undefined) input.autoReview = autoReview;
+
+  // Non-atomic by design: create request first, then resolve issue.
+  // If the second step fails we emit the request id so the operator
+  // knows the partial state and can manually resolve the issue.
+  const req = await c.requestUC.create(input);
+  try {
+    await c.issueUC.setState(id, 'resolved');
+  } catch (e) {
+    const msg = e instanceof Error ? e.message : String(e);
+    process.stderr.write(
+      `⚠ request created but issue state transition failed\n` +
+        `  request: ${req.id.value} (pending)\n` +
+        `  issue:   ${id} (state unchanged)\n` +
+        `  cause:   ${msg}\n` +
+        `  fix:     gate issues resolve ${id}\n`,
+    );
+    return 1;
+  }
+  process.stdout.write(
+    `✓ promoted ${id} → ${req.id.value} (issue resolved)\n`,
+  );
+  return 0;
+}
+
+function resolveIssueVerb(sub: string | undefined): string | undefined {
+  switch (sub) {
+    case 'resolve':
+      return 'resolved';
+    case 'defer':
+      return 'deferred';
+    case 'start':
+      return 'in_progress';
+    case 'reopen':
+      return 'open';
+    default:
+      return undefined;
+  }
+}
